% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_data.R
\name{barplot_legend}
\alias{barplot_legend}
\title{Plot Percentage Legend}
\usage{
barplot_legend(data_plot, color)
}
\arguments{
\item{data_plot}{get it from SDG_network and use positive matrix - SDG_network--"score_pos"-- --"data_pourc"--}

\item{color}{color for each type of NCS}
}
\value{
a the legend of the barplot
}
\description{
Plot Percentage Legend
}
