% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_data.R
\name{barplot_perc_achieve}
\alias{barplot_perc_achieve}
\title{Barplot Percentage Of Target Achieved}
\usage{
barplot_perc_achieve(SDG_network, color, save = FALSE, name)
}
\arguments{
\item{SDG_network}{Matrices with positive and negative relationship - use outputs in SDG_network object}

\item{color}{a character vector specifying the color of each type of NCS}

\item{save}{if TRUE the plot is saved in the results folder}

\item{name}{the name of the plot to be saved}
}
\value{
a barplot with positive and negative values percent of targets linked
}
\description{
Barplot Percentage Of Target Achieved
}
