% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_data.R
\name{barplot_percSDG}
\alias{barplot_percSDG}
\title{Plot Percentage Of Target Achieved}
\usage{
barplot_percSDG(data_plot, save = FALSE)
}
\arguments{
\item{data_plot}{a data frame with percentage of target achieve totally + by group of NCS}

\item{save}{if TRUE the plot is saved in the results folder}
}
\value{
a ggplot object, barplot, of the \% of SDG' target achieved
}
\description{
Plot Percentage Of Target Achieved
}
