% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_data.R
\name{bip_edgewt}
\alias{bip_edgewt}
\title{Scale Edges Size}
\usage{
bip_edgewt(matrix, x)
}
\arguments{
\item{matrix}{a weighted or binary matrix}

\item{x}{a number}
}
\value{
a numeric vector of each edge size scaled as a linear function of the number of link
}
\description{
Scale Edges Size
}
