% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_data.R
\name{biplot_npos_vs_nneg}
\alias{biplot_npos_vs_nneg}
\title{Biplot Of Negative Versus Positive Links}
\usage{
biplot_npos_vs_nneg(data_pos, data_neg, save = TRUE, name)
}
\arguments{
\item{save}{}

\item{name}{}

\item{data}{}
}
\value{

}
\description{
Biplot Of Negative Versus Positive Links
}
