% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_data.R
\name{CA_barplot}
\alias{CA_barplot}
\title{Circular Barplot Of Contribution For Supplementary Fig 1 And 2}
\usage{
CA_barplot(data, axis, variable, ymin, ytitle)
}
\arguments{
\item{data}{obtained with NCSSDGproj::CA_contri_vars 1st element of the list}

\item{axis}{un number corresponding to the axis of the correspondance analysis}

\item{variable}{a character specifying if working on "row" vs "column"}

\item{ymin}{a number for y min value}

\item{ytitle}{a number for title position}

\item{ymax}{a number for y max value}
}
\value{
a circular barplot in Supp Fig 1 and 2
}
\description{
Circular Barplot Of Contribution For Supplementary Fig 1 And 2
}
