% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_data.R
\name{CA_barplot_NCS}
\alias{CA_barplot_NCS}
\title{Barplot Of NCS Contribution}
\usage{
CA_barplot_NCS(data, axis, colNCS_ter, colNCS_coast, colNCS_mar)
}
\arguments{
\item{data}{obrained with NCSSDGproj::CA_contri_vars}

\item{axis}{1 or 2 whether you want to plot 1st or 2nd axis}
}
\value{

}
\description{
Barplot Of NCS Contribution
}
