% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_data.R
\name{CA_barplot_SDG}
\alias{CA_barplot_SDG}
\title{Barplot Of Targets Contribution}
\usage{
CA_barplot_SDG(data_contrib, axis)
}
\arguments{
\item{data_contrib}{}

\item{axis}{}

\item{col_sust}{}

\item{col_unsust}{}

\item{col_oth}{}
}
\value{

}
\description{
Barplot Of Targets Contribution
}
