% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unused_functions.R
\name{CA_contri_vars_unused}
\alias{CA_contri_vars_unused}
\title{Correspondance Analysis Of Variable Contribution}
\usage{
CA_contri_vars_unused(matrix_cont, colNCS_ter, colNCS_coast, colNCS_mar)
}
\arguments{
\item{matrix_cont}{a matrix with NCS in rows and SDG targets in columns - use contingency_mat_targets}

\item{colNCS_ter}{color for terrestrial ecosystems}

\item{colNCS_coast}{color for coastal ecosystems}

\item{colNCS_mar}{color for marine ecosystems}
}
\value{
a list of three files, with CA analysis results, contribution of row to the variance and the data to add arrow on the plot
}
\description{
Correspondance Analysis Of Variable Contribution
}
