% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_data.R
\name{CA_contrib_barplot}
\alias{CA_contrib_barplot}
\title{Separate Correspondance Analysis}
\usage{
CA_contrib_barplot(
  data,
  targ_contrib12,
  NCScontrib12,
  data_arrow,
  colNCS_ter,
  colNCS_coast,
  colNCS_mar,
  save = FALSE
)
}
\arguments{
\item{data}{obtained with NCSSDGproj::CA_contri_vars}

\item{targ_contrib12}{}

\item{NCScontrib12}{}

\item{data_arrow}{}

\item{colNCS_ter}{}

\item{colNCS_coast}{}

\item{colNCS_mar}{}

\item{save}{}
}
\value{

}
\description{
Separate Correspondance Analysis
}
