% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_data.R
\name{CA_contrib_plot}
\alias{CA_contrib_plot}
\title{Title}
\usage{
CA_contrib_plot(data, colNCS_ter, colNCS_coast, colNCS_mar, save = FALSE)
}
\arguments{
\item{data}{}

\item{colNCS_ter}{}

\item{colNCS_coast}{}

\item{colNCS_mar}{}

\item{save}{}
}
\value{

}
\description{
Title
}
