% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_data.R
\name{circular_data_Insurance}
\alias{circular_data_Insurance}
\title{Format Data For Circular Plot}
\usage{
circular_data_Insurance(
  data_Insurance,
  data_long,
  SDG_info,
  NCS_info,
  negative = FALSE
)
}
\arguments{
\item{data_Insurance}{}

\item{data_long}{}

\item{SDG_info}{}

\item{NCS_info}{}

\item{negative}{}
}
\value{

}
\description{
Format Data For Circular Plot
}
