% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_data.R
\name{circular_plot_Insurance}
\alias{circular_plot_Insurance}
\title{Insurance Circular Plot - Figure 3}
\usage{
circular_plot_Insurance(
  data,
  label_data,
  base_data,
  grid_data,
  SDG_info,
  colNCS_ter,
  colNCS_coast,
  colNCS_mar,
  icon_SDG,
  save = FALSE,
  name
)
}
\arguments{
\item{data}{a df with the number of time a target is achieved by each time of ecosystem - use circular_data_Insurance}

\item{label_data}{a df with label text and position - use circular_data_Insurance}

\item{base_data}{a df with the position of each semi-circle - use circular_data_Insurance}

\item{grid_data}{a df with the position of grid - use circular_data_Insurance}

\item{SDG_info}{a dataframe with the category, the color and the name of each SDG and target - use SDG_infos}

\item{colNCS_ter}{color for terrestrial ecosystems}

\item{colNCS_coast}{color for coastal ecosystems}

\item{colNCS_mar}{color for marine ecosystems}

\item{save}{if TRUE the plot is saved in the results folder}

\item{name}{the name of the plot to be saved}

\item{iconSDG}{a list of 16 rastergrob objects for each SDG - use format_icons}
}
\value{
figure 3 in the paper
}
\description{
Insurance Circular Plot - Figure 3
}
