% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_data.R
\name{circular_plot_Insurance_neg}
\alias{circular_plot_Insurance_neg}
\title{Insurance Circular Plot For Negative Data}
\usage{
circular_plot_Insurance_neg(
  data,
  label_data,
  base_data,
  grid_data,
  SDG_info,
  colNCS_ter,
  colNCS_coast,
  colNCS_mar,
  icon_SDG,
  save = FALSE,
  name
)
}
\arguments{
\item{data}{}

\item{label_data}{}

\item{base_data}{}

\item{grid_data}{}

\item{SDG_info}{}

\item{colNCS_ter}{}

\item{colNCS_coast}{}

\item{colNCS_mar}{}

\item{icon_SDG}{}

\item{save}{}

\item{name}{}
}
\value{

}
\description{
Insurance Circular Plot For Negative Data
}
