% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_data.R
\name{contingency_mat_targets}
\alias{contingency_mat_targets}
\title{Contingency Matrix Of SDG's Targets For Network Indices And Unipartit Plot}
\usage{
contingency_mat_targets(raw_dat)
}
\arguments{
\item{raw_dat}{a dataframe with targets of the SDGs in columns and NCSs in rows}
}
\value{
a matrix of 0 and 1
}
\description{
Contingency Matrix Of SDG's Targets For Network Indices And Unipartit Plot
}
