% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_data.R
\name{coords}
\alias{coords}
\title{Railway Coordinates Spread Nodes Equally}
\usage{
coords(mymat, maxX, maxY)
}
\arguments{
\item{mymat}{a weighted or binary matrix with SDG in column and NCS in row}

\item{maxX}{}

\item{maxY}{}
}
\value{
a 2 columns matrix with coordinates of the nodes
}
\description{
Railway Coordinates Spread Nodes Equally
}
