% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{Correspondance_Analysis}
\alias{Correspondance_Analysis}
\title{Correspondance Analysis On Targets}
\usage{
Correspondance_Analysis(matrix01)
}
\arguments{
\item{matrix01}{a matrix with NCS in rows and SDG targets in columns}
}
\value{
a list of two elements, the all CA and CA with subseted variables
}
\description{
Correspondance Analysis On Targets
}
