% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_data.R
\name{data_netw_indice}
\alias{data_netw_indice}
\title{Format Matrix For Network Indice}
\usage{
data_netw_indice(raw_dat)
}
\arguments{
\item{raw_dat}{a dataframe with targets of the SDGs in columns and NCSs in rows}
}
\value{
a matrix of 0 and 1
}
\description{
Format Matrix For Network Indice
}
