% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_data.R
\name{data_TI}
\alias{data_TI}
\title{Format Data For TI Estimate}
\usage{
data_TI(matrix)
}
\arguments{
\item{matrix}{a matrix with targets in rows and NCS in columns - use contingency_mat_targets}
}
\value{
a dataframe with the number of time each target is achieved
}
\description{
Format Data For TI Estimate
}
