% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_data.R
\name{df_to_longDF}
\alias{df_to_longDF}
\title{Data Frame In Long Format}
\usage{
df_to_longDF(df)
}
\arguments{
\item{df}{the raw matrix with targets in columns and NCS in rows - use sheets_to_df}
}
\value{
a df to the long format with score for each SDG's target and each ecosystem
}
\description{
Data Frame In Long Format
}
