% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_data.R
\name{edge_col}
\alias{edge_col}
\title{Set Edges Color}
\usage{
edge_col(matrix)
}
\arguments{
\item{matrix}{a weighted or binary matrix with SDG in column and NCS in row}
}
\value{
edge color for each type of NCS
}
\description{
Set Edges Color
}
