% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_data.R
\name{edge_size}
\alias{edge_size}
\title{Scale Edges Size}
\usage{
edge_size(matrix, x)
}
\arguments{
\item{matrix}{a weighted contingency matrix with SDG in columns and NCS in rows - use matrix_SDG}

\item{x}{a number}
}
\value{
a numeric vector of each edge size scaled as a linear function of the number of link
}
\description{
Scale Edges Size
}
