% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_data.R
\name{Figure2}
\alias{Figure2}
\title{Build Figure Two}
\usage{
Figure2(save = FALSE, name)
}
\arguments{
\item{save}{if TRUE the plot is saved in the results folder}

\item{name}{the name of the plot to be saved}
}
\value{
Figure 2 in the paper
}
\description{
Build Figure Two
}
