% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_data.R
\name{format_icons}
\alias{format_icons}
\title{Icon In Raster Format}
\usage{
format_icons(path, icon_SDG = TRUE)
}
\arguments{
\item{path}{path to load icons}

\item{icon_SDG}{if TRUE, format SDG icons, else, format NCS icons}
}
\value{
A list of raster object with icons ready to plot with ggplot2
}
\description{
Icon In Raster Format
}
