% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_data.R
\name{Insurance_data2plot}
\alias{Insurance_data2plot}
\title{Insurance Data To Plot}
\usage{
Insurance_data2plot(matrix01, Ntarget)
}
\arguments{
\item{matrix01}{a binary matrix with targets in columns and NCS in rows - use contingency_mat_targets}

\item{Ntarget}{the total number of targets linked}
}
\value{
A data frame with number of times a target is achieved with a column identifying observed data vs. null data
}
\description{
Insurance Data To Plot
}
