% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_data.R
\name{Insurance_plot}
\alias{Insurance_plot}
\title{Insurance Plot}
\usage{
Insurance_plot(data, TI, TUI_obs, TUI_null, obs_col, null_col)
}
\arguments{
\item{data}{A data frame with number of times a target is achieved with a column identifying observed data vs. null data}
}
\value{

}
\description{
Insurance Plot
}
