% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_data.R
\name{legend_verti}
\alias{legend_verti}
\title{Produce Vertical Legend}
\usage{
legend_verti(data_plot, color)
}
\arguments{
\item{data_plot}{}

\item{color}{}
}
\value{

}
\description{
Produce Vertical Legend
}
