% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{load_Fig1B}
\alias{load_Fig1B}
\title{Load First Figure Panel B}
\usage{
load_Fig1B()

load_Fig1B()
}
\value{
A RData to plot Figure 1B

A RData to plot Figure 1B
}
\description{
Load First Figure Panel B

Load Legend Of NCS
}
