% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{load_metric_obs}
\alias{load_metric_obs}
\title{Load Observed Metrics}
\usage{
load_metric_obs()
}
\value{
a list of 2 elements, with observed indices values for positive and negative scores
}
\description{
Load Observed Metrics
}
