% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{load_NCS_icon}
\alias{load_NCS_icon}
\title{Load NCS icons path}
\usage{
load_NCS_icon()
}
\value{
a character with the list of paths to read and transform NCS icons
}
\description{
Load NCS icons path
}
