% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{load_SDG_icon}
\alias{load_SDG_icon}
\title{Load SDG icons path}
\usage{
load_SDG_icon()
}
\value{
a character with the list of paths to read and transform SDG icons
}
\description{
Load SDG icons path
}
