% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_data.R
\name{matrix_SDG}
\alias{matrix_SDG}
\title{Weighted Contingency Matrix of SDG}
\usage{
matrix_SDG(data_long)
}
\arguments{
\item{data_long}{A dataframe with 0 and 1 value for each NCS and SDG's targets}
}
\value{
a weighted matrix with SDG in columns and NCS in rows
}
\description{
Weighted Contingency Matrix of SDG
}
