% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unused_functions.R
\name{matrix_SDG_not_use}
\alias{matrix_SDG_not_use}
\title{Weighted Contingency Matrix of SDG - for the 1st scoring system}
\usage{
matrix_SDG_not_use(data_long)
}
\arguments{
\item{data_long}{A dataframe with 0 and 1 value for each NCS and SDG's targets}
}
\value{
a weighted matrix with SDG in columns and NCS in rows
}
\description{
Weighted Contingency Matrix of SDG - for the 1st scoring system
}
