% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_data.R
\name{matrix_to_longDF}
\alias{matrix_to_longDF}
\title{Matrix Long Format}
\usage{
matrix_to_longDF(matrix01)
}
\arguments{
\item{matrix01}{the raw matrix with targets in columns and NCS in rows - use sheets_to_matrix}
}
\value{
a df to the long format with score for each SDG's target and each ecosystem
}
\description{
Matrix Long Format
}
