% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_data.R
\name{matrix_to_network}
\alias{matrix_to_network}
\title{From Matrix to Network object}
\usage{
matrix_to_network(matrix, mode1 = "P", mode2 = "A", pos = TRUE)
}
\arguments{
\item{matrix}{a weighted or binary matrix with SDG in column and NCS in row - use matrix_SDG}

\item{mode1}{}

\item{mode2}{}

\item{pos}{specify positive or negative data}
}
\value{
a network object
}
\description{
From Matrix to Network object
}
