% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_data.R
\name{modularity_plot}
\alias{modularity_plot}
\title{Plot Modularity}
\usage{
modularity_plot(matrix01)
}
\arguments{
\item{matrix01}{formatted matrix to calculate network indices, nestedness and modularity. Use data_netw_indice to format}
}
\value{

}
\description{
Plot Modularity
}
