% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_data.R
\name{NCS_info}
\alias{NCS_info}
\title{Natural Climate Solutions Infos}
\usage{
NCS_info(matrix_cont)
}
\arguments{
\item{matrix_cont}{a matrix with targets in columns and NCS in rows - use contingency_mat_targets}
}
\value{
a dataframe with type of NCS info i.e. terrestrial vs coastal vs marine
}
\description{
Natural Climate Solutions Infos
}
