% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unused_functions.R
\name{network_uniP}
\alias{network_uniP}
\title{Network Format For Unipartite Plot}
\usage{
network_uniP(
  contin_mat_target,
  method,
  colNCS_ter,
  colNCS_coast,
  colNCS_mar,
  colTARG
)
}
\arguments{
\item{contin_mat_target}{a matrix with targets in rows and NCS in columns}

\item{method}{see layout in ggnetwork}

\item{colNCS_ter}{color for terrestrial nodes}

\item{colNCS_coast}{color for coastal nodes}

\item{colNCS_mar}{color for marine nodes}

\item{colTARG}{color for targets nodes}
}
\value{
A data.frame object with coordinates and color columns for each node
}
\description{
Network Format For Unipartite Plot
}
