% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_data.R
\name{null_data_CircPlot}
\alias{null_data_CircPlot}
\title{Null Data For Circular Plot}
\usage{
null_data_CircPlot(matrix01, nsim)
}
\arguments{
\item{matrix01}{matrix with targets in columns and NCS in rows}

\item{nsim}{number of null matrices to be produced}
}
\value{
a data frame with the number of time a target is achieved and the mean value
}
\description{
Null Data For Circular Plot
}
