% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{NullModels}
\alias{NullModels}
\title{Null Model Analysis}
\usage{
NullModels(
  matrix01,
  rawdata,
  NMalgo,
  NESTmethod,
  Nrun,
  Nsim,
  TargetInsurance = FALSE,
  save,
  name
)
}
\arguments{
\item{matrix01}{formatted matrix to calculate network indices, nestedness and modularity - use contingency_mat_targets}

\item{rawdata}{a dataframe with targets of the SDGs in columns and NCSs in rows - use sheets_to_df}

\item{NMalgo}{name of null model algorithm}

\item{NESTmethod}{method to calculate nestedness, use "NODF" or "weighted NODF"}

\item{Nrun}{Number of replicate runs for metaComputesModules}

\item{Nsim}{Number of null matrices}

\item{TargetInsurance}{if TRUE, compute Insurance indices, if FALSE calculate  network indices}

\item{save}{if statement to save the results}

\item{name}{the name of the data to be saved}
}
\value{
a 3 column dataframe with null model results for nestedness and modularity OR TOi and TUI
}
\description{
Null Model Analysis
}
