% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_data.R
\name{perc_SDG}
\alias{perc_SDG}
\title{Percentage Of Target Achieved}
\usage{
perc_SDG(data_long)
}
\arguments{
\item{data_long}{a dataframe to the long format with value 0 or 1 for each SDG's target and each ecosystem}
}
\value{
a data frame with percentage of target achieve totally + by group of NCS, values to be plotted
}
\description{
Percentage Of Target Achieved
}
