% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_data.R
\name{percentage_of_ones}
\alias{percentage_of_ones}
\title{Percentage Of Ones In Each Ecosystem}
\usage{
percentage_of_ones(data_pos, data_neg, save = TRUE, name)
}
\arguments{
\item{data_pos}{matrix with positive values}

\item{data_neg}{matrix with negative values}

\item{save}{if statement, if TRUE the plot is saved}

\item{name}{the name of the plot to be saved}
}
\value{

}
\description{
Percentage Of Ones In Each Ecosystem
}
