% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unused_functions.R
\name{plot_CorresAna}
\alias{plot_CorresAna}
\title{Correspondance Analysis Plot}
\usage{
plot_CorresAna(
  ca,
  ca_subset,
  NCS_info,
  colors,
  ellipse = TRUE,
  hull = FALSE,
  save = FALSE
)
}
\arguments{
\item{ca}{matrix or data frame of axis scores for each observation obtained with ade4::dudi.coa}

\item{ca_subset}{a susbet of variables}

\item{NCS_info}{dataframe obtain with NCSSDGproj::NCS_info}

\item{colors}{3 colors vector for terrestrial, coastal and marine NCS}

\item{ellipse}{show ellipse default = TRUE}

\item{hull}{show convexhull default = FALSE}

\item{save}{save plot default = FALSE}
}
\value{
a CA plot
}
\description{
Correspondance Analysis Plot
}
