% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_data.R
\name{plot_n_links}
\alias{plot_n_links}
\title{Plot The Number Of Links By Ecosystems}
\usage{
plot_n_links(data_pos, data_neg, save = TRUE, name)
}
\arguments{
\item{save}{}

\item{name}{}

\item{data}{}
}
\value{

}
\description{
Plot The Number Of Links By Ecosystems
}
