% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_data.R, R/unused_functions.R
\name{plot_network}
\alias{plot_network}
\title{Plot Network As Flux Diagramm}
\usage{
plot_network(network_obj, matrix, icon_SDG, icon_NCS, nodes_col, save = FALSE)

plot_network(network_obj, matrix, icon_SDG, icon_NCS, nodes_col, save = FALSE)
}
\arguments{
\item{network_obj}{}

\item{matrix}{}

\item{icon_SDG}{}

\item{icon_NCS}{}
}
\value{
a flux diagramm of the links between SDG and NCS


}
\description{
Plot Network As Flux Diagramm

Plot Network - for the 1st scoring system
}
