% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unused_functions.R
\name{plot_network_unused}
\alias{plot_network_unused}
\title{Plot Network - for the 1st scoring system}
\usage{
plot_network_unused(
  network_obj,
  matrix,
  icon_SDG,
  icon_NCS,
  nodes_col,
  save = FALSE
)
}
\arguments{
\item{network_obj}{}

\item{matrix}{}

\item{icon_SDG}{}

\item{icon_NCS}{}
}
\value{

}
\description{
Plot Network - for the 1st scoring system
}
