% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_data.R
\name{plot_network_V2}
\alias{plot_network_V2}
\title{Plot Network Version Two}
\usage{
plot_network_V2(
  network_obj,
  matrix,
  icon_SDG,
  icon_NCS,
  nodes_col,
  save = FALSE
)
}
\arguments{
\item{network_obj}{}

\item{matrix}{}

\item{icon_SDG}{}

\item{icon_NCS}{}
}
\value{

}
\description{
Plot Network Version Two
}
