% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{read_all_sheets}
\alias{read_all_sheets}
\title{Read All Sheets Of The Supplementary Material}
\usage{
read_all_sheets()
}
\value{
a list of dataframe for each ecosystem
}
\description{
Read All Sheets Of The Supplementary Material
}
