% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{SDG_contrib_tbl}
\alias{SDG_contrib_tbl}
\title{Read the binary matrix}
\usage{
SDG_contrib_tbl()
}
\value{
a dataframe with target's contribution and type to the second axis variance
}
\description{
Read the binary matrix
}
