% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_data.R
\name{SDG_infos}
\alias{SDG_infos}
\title{Sustainable Development Goals Infos}
\usage{
SDG_infos(matrix_cont)
}
\arguments{
\item{matrix_cont}{a matrix with targets in columns and NCS in rows}
}
\value{
a dataframe with with the category, the color and the name of each SDG and target
}
\description{
Sustainable Development Goals Infos
}
