% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{SES_pval}
\alias{SES_pval}
\title{SES Pvalue Calculation}
\usage{
SES_pval(val_obs, mean_null, sd_null, rowname)
}
\arguments{
\item{val_obs}{observed value of the indice}

\item{mean_null}{average value of the indice computed on null matrices}

\item{sd_null}{sd value of the indice computed on null matrices}

\item{rowname}{a character specifying rowname}
}
\value{
return a 3 columns dataframe with null model results, SES and pvalues
}
\description{
SES Pvalue Calculation
}
