% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_data.R
\name{sheets_to_matrix}
\alias{sheets_to_matrix}
\title{Transform Sheets From Each Ecosystem Into Combined Matrix}
\usage{
sheets_to_matrix(sheets_list)
}
\arguments{
\item{sheets_list}{a list of dataframes obtained}
}
\value{
a list of two elements with positive and negative scores matrix
}
\description{
Transform Sheets From Each Ecosystem Into Combined Matrix
}
