% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_data.R
\name{supp_fig1}
\alias{supp_fig1}
\title{Supplementary Fig 1}
\usage{
supp_fig1(
  data,
  targ_contrib12,
  data_arrow,
  colNCS_ter,
  colNCS_coast,
  colNCS_mar,
  save = FALSE,
  name
)
}
\arguments{
\item{data}{obtained with NCSSDGproj::CA_contri_vars}

\item{targ_contrib12}{}

\item{data_arrow}{}

\item{colNCS_ter}{}

\item{colNCS_coast}{}

\item{colNCS_mar}{}

\item{save}{}

\item{name}{the name of the plot to be saved}
}
\value{

}
\description{
Supplementary Fig 1
}
