% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_data.R
\name{supp_fig3_4}
\alias{supp_fig3_4}
\title{Biplot Of Negative Versus Positive Links}
\usage{
supp_fig3_4(data_pos, data_neg, save = TRUE, name1, biplot = TRUE, name2)
}
\arguments{
\item{save}{if TRUE the plot is saved in the results folder}

\item{data}{data to be plotted - use sheets_to_df}

\item{name}{the name of the plot to be saved}
}
\value{
Supp figure 3 and 4 in the paper
}
\description{
Biplot Of Negative Versus Positive Links
}
