% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_data.R
\name{supp_fig5to8}
\alias{supp_fig5to8}
\title{Plot Null Values Vs. Observed Value With Percentiles}
\usage{
supp_fig5to8(null_vals, res_null_mod, name)
}
\arguments{
\item{null_vals}{a list of null values for modularity, nestedness, TUI and TOI}

\item{res_null_mod}{a dataframe with 2.5 and 97.5 percentiles and observed values for modularity, nestedness, TUI and TOI - use load_metric_obs}

\item{name}{the name of the plot to be saved}
}
\value{

}
\description{
Plot Null Values Vs. Observed Value With Percentiles
}
