% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{TUI_TOI}
\alias{TUI_TOI}
\title{Target Under Insurance -TUI- And Over Insurance -TOI-
Functional indices adapted for Mouillot et al. 2014 - Functional over-redundancy and high functional vulnerability in global fish faunas on tropical reefs - https://doi.org/10.1073/pnas.1317625111}
\usage{
TUI_TOI(data_TI, Necosystem, Ntarget)
}
\arguments{
\item{data_TI}{a dataframe with the number of time each target is achieved}

\item{Necosystem}{the number of ecosystem}

\item{Ntarget}{the number of targets}
}
\value{
a three columns dataframe with TUI, TI and TOI values
}
\description{
Target Under Insurance -TUI- And Over Insurance -TOI-
Functional indices adapted for Mouillot et al. 2014 - Functional over-redundancy and high functional vulnerability in global fish faunas on tropical reefs - https://doi.org/10.1073/pnas.1317625111
}
