% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{turn_values_randomly}
\alias{turn_values_randomly}
\title{Randomly Turn Values In A Matrix}
\usage{
turn_values_randomly(data_links, percentage)
}
\arguments{
\item{data_links}{a dataframe with links bewteen targets -in columns- and NCS -in rows-}

\item{percentage}{percentage of values that match to be replaced}
}
\value{
a dataframe with modified values
}
\description{
Randomly Turn Values In A Matrix
}
