% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_data.R
\name{unipart_plot}
\alias{unipart_plot}
\title{Unipartite Plot Of Targets}
\usage{
unipart_plot(netw, colNCS_ter, colNCS_coast, colNCS_mar, save)
}
\arguments{
\item{netw}{a dataframe network object from the network_uniP function}

\item{colNCS_ter}{color for terrestrial nodes}

\item{colNCS_coast}{color for coastal nodes}

\item{colNCS_mar}{color for marine nodes}

\item{save}{if TRUE the plot is saved in the results folder}
}
\value{

}
\description{
Unipartite Plot Of Targets
}
