% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_data.R
\name{barplot_perc_achieve}
\alias{barplot_perc_achieve}
\title{Barplot Percentage Of Target Achieved}
\usage{
barplot_perc_achieve(SDG_network, color, save = FALSE)
}
\arguments{
\item{SDG_network}{Matrices with positive and negative relationship - use outputs SDG_network}

\item{color}{color for each type of NCS}

\item{save}{if statement to save the plot}
}
\value{
a barplot with positive and negative values from targets achievement
}
\description{
Barplot Percentage Of Target Achieved
}
