% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_data.R
\name{CA_barplot}
\alias{CA_barplot}
\title{Circular Barplot Of Contribution For Supplementary Fig 1 And 2}
\usage{
CA_barplot(data, axis, variable, ymin, ytitle)
}
\arguments{
\item{data}{obtained with NCSSDGproj::CA_contri_vars 1st element of the list}

\item{axis}{}

\item{variable}{}

\item{ymin}{}

\item{ytitle}{}

\item{ymax}{}
}
\value{

}
\description{
Circular Barplot Of Contribution For Supplementary Fig 1 And 2
}
