% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_data.R
\name{CA_barplot}
\alias{CA_barplot}
\title{Barplot Of Contribution}
\usage{
CA_barplot(data, axis, variable, TOP20, ymin, ymax, ytitle)
}
\arguments{
\item{data}{obtained with NCSSDGproj::CA_contri_vars 1st element of the list}

\item{axis}{}

\item{col_sust}{}

\item{col_unsust}{}

\item{col_oth}{}
}
\value{

}
\description{
Barplot Of Contribution
}
