% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{CA_contri_vars}
\alias{CA_contri_vars}
\title{Correspondance Analysis Of Variable Contribution}
\usage{
CA_contri_vars(matrix01, colNCS_ter, colNCS_coast, colNCS_mar)
}
\arguments{
\item{matrix01}{a matrix with NCS in rows and SDG targets in columns}
}
\value{

}
\description{
Correspondance Analysis Of Variable Contribution
}
