% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_data.R
\name{circular_data_CA}
\alias{circular_data_CA}
\title{Data To Plot Circular Contribution}
\usage{
circular_data_CA(data_contrib, variable, TOP20, axis)
}
\arguments{
\item{data_contrib}{obtained with NCSSDGproj::CA_contri_vars}

\item{variable}{format contribution of ROW of COLUMN - character row or col}

\item{TOP20}{use if variable = "col"}

\item{axis}{axis to be plotted}
}
\value{

}
\description{
Data To Plot Circular Contribution
}
