% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_data.R
\name{circular_data_Insurance}
\alias{circular_data_Insurance}
\title{Title}
\usage{
circular_data_Insurance(data_Insurance, SDG_info)
}
\arguments{
\item{data_Insurance}{}

\item{SDG_info}{}
}
\value{

}
\description{
Title
}
