% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_data.R
\name{circular_plot_Insurance}
\alias{circular_plot_Insurance}
\title{Insurance Circular Plot}
\usage{
circular_plot_Insurance(data, label_data, base_data, grid_data, SDG_info)
}
\arguments{
\item{data}{}

\item{label_data}{}

\item{base_data}{}

\item{grid_data}{}
}
\value{

}
\description{
Insurance Circular Plot
}
