% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_data.R
\name{coords}
\alias{coords}
\title{Railway Coordinates Spread Nodes Equally}
\usage{
coords(mymat, maxX, maxY)
}
\arguments{
\item{mymat}{a weighted contingency matrix with SDG in columns and NCS in rows - use matrix_SDG}

\item{maxX}{a number}

\item{maxY}{a number}
}
\value{
a 2 columns matrix with coordinates of each nodes
}
\description{
Railway Coordinates Spread Nodes Equally
}
