% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_data.R
\name{Figure3}
\alias{Figure3}
\title{Figure 3}
\usage{
Figure3(
  data,
  targ_contrib12,
  data_arrow,
  TOP20_1,
  TOP20_2,
  colNCS_ter,
  colNCS_coast,
  colNCS_mar,
  save = FALSE
)
}
\arguments{
\item{data}{obtained with NCSSDGproj::CA_contri_vars}

\item{targ_contrib12}{}

\item{data_arrow}{}

\item{colNCS_ter}{}

\item{colNCS_coast}{}

\item{colNCS_mar}{}

\item{save}{}

\item{NCScontrib12}{}
}
\value{

}
\description{
Figure 3
}
