% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_data.R
\name{Insurance_data2plot}
\alias{Insurance_data2plot}
\title{Insurance Data To Plot}
\usage{
Insurance_data2plot(matrix01, Ntarget)
}
\arguments{
\item{matrix01}{a matrix with targets in rows and NCS in columns}

\item{Ntarget}{the total number of targets}
}
\value{
A data frame with number of times a target is achieved with a column identifying observed data vs. null data
}
\description{
Insurance Data To Plot
}
