% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_data.R
\name{Insurance_plot}
\alias{Insurance_plot}
\title{Insurance Plot}
\usage{
Insurance_plot(data, TI, TUI_obs, TUI_null, obs_col, null_col, save)
}
\arguments{
\item{data}{A data frame with number of times a target is achieved with a column identifying observed data vs. null data}

\item{TI}{Target Insurance}

\item{TUI_obs}{Target Under Insurance observed}

\item{TUI_null}{Target Under Insurance from null matrix}

\item{obs_col}{color for observed data}

\item{null_col}{color for null data}

\item{save}{if TRUE the plot is saved in the results folder}
}
\value{

}
\description{
Insurance Plot
}
