% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_data.R
\name{matrix_SDG}
\alias{matrix_SDG}
\title{Weighted Contingency Matrix of SDG}
\usage{
matrix_SDG(data_long)
}
\arguments{
\item{data_long}{A dataframe with score for each NCS and SDG's targets - use df_to_longDF}
}
\value{
a weighted matrix of links between SDG in columns and NCS in rows
}
\description{
Weighted Contingency Matrix of SDG
}
