% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{NullModels}
\alias{NullModels}
\title{Null Model Analysis}
\usage{
NullModels(
  matrix01,
  rawdata,
  NMalgo,
  NESTmethod = c("NODF", "weighted NODF"),
  Nrun,
  TargetInsurance = FALSE
)
}
\arguments{
\item{matrix01}{formatted matrix to calculate network indices, nestedness and modularity}

\item{rawdata}{a dataframe with targets of the SDGs in columns and NCSs in rows}

\item{NMalgo}{name of null model algorithm}

\item{NESTmethod}{"NODF" or "weighted NODF"}

\item{Nrun}{Number of replicate runs}

\item{TargetInsurance}{if TRUE, compute Insurance indices, not network indices}
}
\value{

}
\description{
Null Model Analysis
}
