% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_data.R
\name{plot_network}
\alias{plot_network}
\title{Plot Network}
\usage{
plot_network(network_obj, matrix, icon_SDG, icon_NCS)
}
\arguments{
\item{network_obj}{}

\item{matrix}{}

\item{icon_SDG}{}

\item{icon_NCS}{}
}
\value{

}
\description{
Plot Network
}
