% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_data.R
\name{plot_network_neg}
\alias{plot_network_neg}
\title{Plot Network As Flux Diagramm For Negative Links}
\usage{
plot_network_neg(
  network_obj,
  matrix,
  icon_SDG,
  icon_NCS,
  nodes_col,
  save = FALSE,
  name
)
}
\arguments{
\item{network_obj}{a network object - use matrix_to_network}

\item{matrix}{a weighted contingency matrix with SDG in columns and NCS in rows - use matrix_SDG}

\item{icon_SDG}{a list of 16 rastergrob objects for each SDG - use format_icons}

\item{icon_NCS}{a list of 11 rastergrob objects for each NCS - use format_icons}

\item{nodes_col}{}

\item{save}{}

\item{name}{}
}
\value{
a flux diagramm of the links between SDG and NCS
}
\description{
Plot Network As Flux Diagramm For Negative Links
}
