% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{sensitivity_analysis}
\alias{sensitivity_analysis}
\title{Run Sensitivity Analysis}
\usage{
sensitivity_analysis(matrix_rep, obs_values, Nrun, save = TRUE, name)
}
\arguments{
\item{matrix_rep}{a list of matrix, use contingency_mat_targets}

\item{obs_values}{observed metric}

\item{Nrun}{Number of replicate runs for metaComputesModules}

\item{save}{if statement to save the results}

\item{name}{the name of the data to be saved}
}
\value{
a dataframe with observed and modify metrics and SES and pvalue
}
\description{
Run Sensitivity Analysis
}
