% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_data.R
\name{sheets_to_matrix}
\alias{sheets_to_matrix}
\title{Transform Sheets From Each Ecosystem Into Combined Matrix}
\usage{
sheets_to_matrix(sheets_list)
}
\arguments{
\item{sheets_list}{a list of dataframes obtained with read_all_sheets}
}
\value{
a list of four elements -df- with positive, negative, net and cumulate scores matrix
}
\description{
Transform Sheets From Each Ecosystem Into Combined Matrix
}
