% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_data.R
\name{supp_fig1}
\alias{supp_fig1}
\title{Supplementary Fig 1}
\usage{
supp_fig1(
  data,
  arrow,
  data_arrow,
  colNCS_ter,
  colNCS_coast,
  colNCS_mar,
  save = FALSE,
  name
)
}
\arguments{
\item{data}{obtained with NCSSDGproj::CA_contri_vars}

\item{arrow}{TRUE if arrow must be plotted}

\item{data_arrow}{if TRUE, a df specifying the position of each arrow}

\item{colNCS_ter}{color for terrestrial ecosystems}

\item{colNCS_coast}{color for coastal ecosystems}

\item{colNCS_mar}{color for marine ecosystems}

\item{save}{if TRUE the plot is saved in the results folder}

\item{name}{the name of the plot to be saved}
}
\value{
the Supplementary Fig 1 in the paper
}
\description{
Supplementary Fig 1
}
