% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_data.R
\name{supp_fig3_4}
\alias{supp_fig3_4}
\title{Biplot Of Negative Versus Positive Links}
\usage{
supp_fig3_4(data_pos, data_neg, save = TRUE, name1, biplot = TRUE, name2)
}
\arguments{
\item{save}{}

\item{data}{}

\item{name}{}
}
\value{

}
\description{
Biplot Of Negative Versus Positive Links
}
