% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{TUI_TOI}
\alias{TUI_TOI}
\title{Target Under Insurance And Over Insurance}
\usage{
TUI_TOI(data_TI, Necosystem, Ntarget)
}
\arguments{
\item{data_TI}{a dataframe with the number of time each target is achieved}

\item{Necosystem}{the number of ecosystem}

\item{Ntarget}{the number of targets}
}
\value{
a three columns dataframe with TUI, TI and TOI values
}
\description{
Target Under Insurance And Over Insurance
}
